Moving functions to their appropriate files
