+++ src/exceptions.py
@@ -1,0 +1,3 @@
+class BillyError(Exception):
+    """Custom exception for Billy application errors."""
+    pass
